Code quality: remove dead usings/casts, add modifiers, readonly fields
Auto-applied with 'dotnet format' and verified (build green, tests pass):
- IDE0005: remove unnecessary using directives
- IDE0004: remove unnecessary casts (all identity/implicit conversions)
- IDE0040: add explicit accessibility modifiers matching the existing default
- IDE0044: mark fields readonly where never reassigned
- IDE1005: use conditional delegate invocation (?.Invoke) for events

IDE0059 (unnecessary assignment) was deliberately excluded: its fixer
rewrites named, commented reads into bare '_ =' discards and misfired on
interop structs, so it stays at suggestion rather than being auto-applied.

+++ src/NAudio.Alsa/AlsaPcm.cs
@@ -110,7 +110,7 @@ internal sealed class AlsaPcm : IDisposable
                     AlsaInterop.PcmHwParamsSetPeriodsNear(Pcm, hw, ref periods, ref dir),
                     "snd_pcm_hw_params_set_periods_near");
 
-                nuint bufferFrames = (nuint)(PeriodFrames * Periods);
+                nuint bufferFrames = PeriodFrames * Periods;
                 AlsaException.ThrowIfError(
                     AlsaInterop.PcmHwParamsSetBufferSizeNear(Pcm, hw, ref bufferFrames),
                     "snd_pcm_hw_params_set_buffer_size_near");
